aws_bedrock/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
Flask>=3.0.0
flask-cors>=4.0.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
